+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 3. [](#bugfix)
     * Fix collections support in BluprintSchema::extra()
     * Fix exception in stream wrapper when scheme is not defined in locator 
+    * Prevent UniformResourceLocator from resolving paths outside of defined scheme paths (#8)
 
 # v1.3.0
 ## 03/07/2016