+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Fix collections support in BluprintSchema::extra()
     * Fix exception in stream wrapper when scheme is not defined in locator 
     * Prevent UniformResourceLocator from resolving paths outside of defined scheme paths (#8)
+    * Fix breaking YAML files which start with three dashes (#5) 
 
 # v1.3.0
 ## 03/07/2016